% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.prices.R
\name{get.prices}
\alias{get.prices}
\title{get.prices}
\usage{
get.prices(symbols = c("AAPL", "IBM", "XOM"),
  from = as.Date("2014-11-01"), to = Sys.Date())
}
\arguments{
\item{symbols}{a vector of ticker symbols}

\item{from}{date of first price}

\item{to}{date of last price}
}
\value{
an environment of OHLC data
}
\description{
Retrieve prices of a vector of symbols from the web (yahoo)
}
