\name{iSpeculator-package}
\alias{iSpeculator-package}
\alias{iSpeculator}
\docType{package}
\title{
\packageTitle{iSpeculator}
}
\description{
\packageDescription{iSpeculator}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iSpeculator}
\packageIndices{iSpeculator}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{iSpeculator}

Maintainer: \packageMaintainer{iSpeculator}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
